refactor: remove unused SCSS files and styles
- Deleted Transactions.module.scss, _resets.scss, _typography.scss, _button.scss, _variables.scss, _normalize.scss, and vercel.json as they are no longer needed.
- Removed vite.config.js and replaced it with a new vite.config.ts file for TypeScript support.
- Added new TypeScript configuration files: tsconfig.app.json, tsconfig.node.json, and tsconfig.json for better type checking and module resolution.
- Created new main.tsx and App.tsx files to set up the main application structure.
- Implemented new pages: Budget, Login, Overview, Pots, RecurringBills, Register, and Transactions with their respective styles.
- Updated the project to use React Router for navigation and hooks for state management.
