+++ src/App.tsx
@@ -1,0 +1,5 @@
+function App() {
+  return <div>App</div>;
+}
+
+export default App;
