--- src/index.scss
@@ -1,5 +1,0 @@
-@import "./scss/utils/variables";
-@import "./scss/vendors/normalize";
-@import "./scss/base/typography";
-@import "./scss/base/resets";
-@import "./scss/components/button";
