--- src/main.jsx
@@ -1,12 +1,0 @@
-import { createRoot } from "react-dom/client";
-import "./index.scss";
-import App from "./App.jsx";
-
-import { store } from "./app/store.js";
-import { Provider } from "react-redux";
-
-createRoot(document.getElementById("root")).render(
-  <Provider store={store}>
-    <App />
-  </Provider>
-);
