+++ src/pages/Login.jsx
@@ -1,27 +1,24 @@
-import "./Register.scss";
+import "./Login.scss";
 
-import { useRegister } from "../../hooks/useRegister";
-
-import FormInput from "../../components/formInput/FormInput";
+import FormInput from "../components/formInput/FormInput";
 import { Link } from "react-router-dom";
+import { useLogin } from "../hooks/useLogin";
 
-function Register() {
-  const { user, isPending, register } = useRegister();
-
-  const handleRegister = (e) => {
+function Login() {
+  const { login, isPending } = useLogin();
+  const handleLogin = (e) => {
     e.preventDefault();
     const formData = new FormData(e.target);
 
     const email = formData.get("email");
-    const displayName = formData.get("displayName");
     const password = formData.get("password");
 
-    register(email, password, displayName);
+    login(email, password);
   };
   return (
     <div className="login-page">
       <div className="login-left">
-        <img src="./images/logo-large.svg" alt="register image" />
+        <img src="./images/logo-large.svg" alt="login image" />
         <div>
           <h5 className="login-left-title">
             Keep track of your money and save for your future
@@ -38,14 +35,8 @@ function Register() {
         alt="finance image"
       />
       <div className="login-form">
-        <h1 className="login-title">Sign Up</h1>
-        <form onSubmit={handleRegister}>
-          <FormInput
-            label="Name"
-            name="displayName"
-            placeholder="Type here..."
-            type="text"
-          />
+        <h1 className="login-title">Login</h1>
+        <form onSubmit={handleLogin}>
           <FormInput
             label="Email"
             name="email"
@@ -53,19 +44,19 @@ function Register() {
             type="email"
           />
           <FormInput
-            label="Create Password"
+            label="Password"
             name="password"
             placeholder="Type here..."
             type="password"
           />
           <button className="login-btn">
-            {isPending ? "Loading..." : "Create Account"}
+            {isPending ? "Loading..." : "Login"}
           </button>
         </form>
         <p className="register-caption">
-          Already have an account?
-          <Link className="link-register" to="/login">
-            Login
+          Need to create account?{" "}
+          <Link className="link-register" to="/register">
+            SignUp
           </Link>
         </p>
       </div>
@@ -73,4 +64,4 @@ function Register() {
   );
 }
 
-export default Register;
+export default Login;
