+++ src/pages/index.js
@@ -1,7 +1,7 @@
-export { default as Budget } from "./budgets/Budget";
-export { default as Login } from "./login/Login";
-export { default as Overview } from "./overview/Overview";
-export { default as Pots } from "./pots/Pots";
-export { default as RecurringBills } from "./recurringBills/RecurringBills";
-export { default as Register } from "./register/Register";
-export { default as Transactions } from "./transactions/Transactions";
+// export { default as Budget } from "./Budget";
+export { default as Login } from "./Login";
+export { default as Overview } from "./Overview";
+export { default as Pots } from "./Pots";
+export { default as RecurringBills } from "./RecurringBills";
+export { default as Register } from "./Register";
+export { default as Transactions } from "./Transactions";
