--- src/pages/login/Login.scss
@@ -1,101 +1,0 @@
-.login-page {
-  display: flex;
-  align-items: center;
-  gap: 140px;
-  margin-right: auto;
-  margin-left: auto;
-  position: relative;
-}
-
-.login-title {
-  text-align: left;
-  margin: 0;
-  padding: 0;
-}
-
-.login-image {
-  position: absolute;
-  z-index: -5;
-  top: 20px;
-  left: 20px;
-  border-radius: 12px;
-  height: 700px;
-  width: 550px;
-}
-
-.login-left {
-  width: 550px;
-  height: 700px;
-  padding: $spacing-500;
-  margin: 20px;
-  border-radius: 12px;
-  display: flex;
-  justify-content: space-between;
-  flex-direction: column;
-  img {
-    width: 121px;
-    height: 21px;
-  }
-}
-
-.login-left-title {
-  color: $white;
-  font-weight: 700;
-  font-size: $text-preset-1;
-  margin-bottom: 24px;
-}
-
-.login-left-caption {
-  color: $white;
-  font-size: $text-preset-4;
-  line-height: 150%;
-}
-
-.login-form {
-  display: flex;
-  flex-direction: column;
-  //   align-items: center;
-  gap: 32px;
-  background-color: $white;
-  padding: $spacing-400;
-  border-radius: $spacing-150;
-  form {
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-  }
-}
-
-.register-caption {
-  text-align: center;
-  color: $grey-500;
-  font-size: $text-preset-4;
-  .link-register {
-    color: $grey-900;
-    font-weight: 700;
-    font-size: $text-preset-4;
-    text-decoration: underline;
-  }
-}
-
-.login-btn {
-  padding: $spacing-200;
-  background-color: $grey-900;
-  border-radius: $spacing-100;
-  color: $white;
-  margin-top: 32px;
-  cursor: pointer;
-}
-
-@media (max-width: 1210px) {
-  .login-left {
-    display: none !important;
-    opacity: 50%;
-    // visibility: hidden;
-  }
-  .login-page {
-    height: 100%;
-    align-items: center;
-    justify-content: center;
-  }
-}
