--- src/pages/overview/Overview.module.scss
@@ -1,344 +1,0 @@
-.overview {
-  display: flex;
-  flex-direction: column;
-  gap: $spacing-400;
-  padding: $spacing-500 $spacing-400;
-  height: 700px;
-}
-
-.overview-title {
-  font-size: $text-preset-1;
-  font-weight: 700;
-  line-height: $line-height-120;
-  color: $grey-900;
-  margin: 0;
-}
-
-.overview-top {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  gap: $spacing-300;
-  height: 119px;
-  div:nth-child(1) {
-    background-color: $grey-900;
-    span,
-    h6 {
-      color: white;
-    }
-  }
-  div {
-    background-color: $white;
-    border-radius: $spacing-150;
-    padding: $spacing-300;
-    span {
-      color: $grey-500;
-      font-size: $text-preset-4;
-      line-height: $line-height-150;
-    }
-    h6 {
-      font-weight: 700;
-      font-size: $text-preset-1;
-      line-height: $line-height-120;
-      margin-top: 12px;
-    }
-  }
-}
-
-.overview-bottom {
-  display: grid;
-  grid-template-columns: 1.5fr 1fr;
-  gap: 24px;
-  height: 1060px;
-}
-
-.overview-pots {
-  background: $white;
-  padding: $spacing-400;
-  border-radius: $spacing-150;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.overview-pots-top {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  h5 {
-    font-weight: 700;
-    font-size: $text-preset-2;
-    line-height: $line-height-120;
-    color: $grey-900;
-  }
-  a {
-    color: $grey-500;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-    display: inline-flex;
-    gap: 12px;
-    align-items: center;
-  }
-}
-
-.overview-pots-bottom {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 20px;
-  .overview-pots-bottom-left {
-    background-color: $beige-100;
-    border-radius: $spacing-150;
-    padding: $spacing-200;
-    display: flex;
-    align-items: center;
-    gap: 16px;
-    .caption {
-      color: $grey-500;
-      font-size: $text-preset-4;
-      line-height: $line-height-150;
-    }
-    .wrapper {
-      display: flex;
-      flex-direction: column;
-      gap: 11px;
-    }
-  }
-  svg {
-    height: 40px;
-    width: 40px;
-  }
-}
-
-.overview-budgets {
-  height: 410px;
-  background-color: $white;
-  grid-column: 2;
-  grid-row: 1 / 3;
-  border-radius: $spacing-150;
-  padding: $spacing-400;
-}
-
-.overview-recurring {
-  background-color: $white;
-  grid-column: 2;
-  grid-row: 3 / 4;
-  border-radius: $spacing-150;
-  padding: $spacing-400;
-}
-
-.overview-bottom_wrappers {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.overview-pots-bottom_title {
-  font-weight: 700;
-  color: $grey-900;
-  line-height: $line-height-120;
-  font-size: $text-preset-1;
-}
-
-.overview-pots-bottom_list {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 16px;
-}
-
-.overview-pots-bottom_list-item {
-  display: flex;
-  div span {
-    color: $grey-500;
-    font-size: $text-preset-5;
-    line-height: $line-height-150;
-  }
-  div h6 {
-    font-weight: 700;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-    color: $grey-900;
-  }
-}
-
-.overview-pots-bottom_list-item_shape {
-  width: 4px;
-  height: 43px;
-  border-radius: $spacing-100;
-  display: inline-block;
-  margin-right: 16px;
-}
-
-.overview-transactions {
-  background: white;
-  border-radius: $spacing-150;
-  padding: $spacing-400;
-  display: flex;
-  flex-direction: column;
-  gap: $spacing-400;
-}
-
-.overview-transactions-top {
-  display: flex;
-  justify-content: space-between;
-  h5 {
-    font-size: $text-preset-2;
-    font-weight: 700;
-    line-height: $line-height-120;
-    color: $grey-900;
-  }
-  a {
-    display: inline-flex;
-    align-items: center;
-    gap: 16px;
-    color: $grey-500;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-  }
-}
-
-.overview-transactions-list {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.overview-transaction {
-  display: flex;
-  justify-content: space-between;
-  padding-bottom: 20px;
-  border-bottom: 1px $grey-100 solid;
-}
-
-.overview-transaction-user {
-  display: flex;
-  gap: 16px;
-  align-items: center;
-  img {
-    height: 40px;
-    width: 40px;
-    border-radius: 50%;
-  }
-}
-
-.overview-transaction-captions {
-  display: flex;
-  flex-direction: column;
-  text-align: right;
-  h6 {
-    font-weight: 700;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-  }
-  span {
-    color: $grey-500;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-  }
-}
-
-.overview-section-top {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-  h5 {
-    font-weight: 700;
-    font-size: $text-preset-2;
-    line-height: $line-height-120;
-    color: $grey-900;
-  }
-  a {
-    display: inline-flex;
-    gap: 12px;
-    align-items: center;
-    color: $grey-500;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-  }
-}
-
-.overview-budgets-chart {
-  width: 247px;
-  height: 239px;
-
-  color: white;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.overview-budgets_bottom {
-  display: flex;
-  gap: 16px;
-}
-
-.overview-budgets-list {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.overview-budgets-list_item {
-  span {
-    border-radius: $spacing-50;
-    width: 4px;
-    background-color: red;
-  }
-  div span {
-    font-size: $text-preset-5;
-    line-height: $line-height-150;
-    color: $grey-500;
-  }
-  h6 {
-    font-weight: 700;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-    color: $grey-900;
-  }
-}
-
-.overview-recurring__list {
-  margin-top: 32px;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.overview-recurring__list-item {
-  padding: 16px 20px;
-  border-radius: $spacing-100;
-  border-left: 4px solid transparent;
-  background-color: $beige-100;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  span {
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-    color: $grey-500;
-  }
-  h5 {
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-    color: $grey-900;
-  }
-}
-
-@media (max-width: 1020px) {
-  .overview-top {
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-  }
-}
-
-@media (max-width: 768px) {
-  .overview-top {
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-  }
-}
-
-.active-transaction {
-  color: $green;
-}
-
-.passive-transaction {
-  color: $grey-900;
-}
