--- src/pages/transactions/Transactions.module.scss
@@ -1,142 +1,0 @@
-.transactions-page {
-  padding: 40px;
-  max-height: 99dvh;
-}
-
-.transactions-title {
-  margin-top: 1px;
-  font-size: $text-preset-1;
-  line-height: $line-height-120;
-  color: $grey-900;
-  margin-bottom: 32px;
-}
-
-.transactions {
-  background-color: $white;
-  padding: $spacing-400;
-  border-radius: $spacing-150;
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.transactions-top {
-  display: flex;
-  justify-content: space-between;
-}
-
-.transactions-search {
-  display: flex;
-  gap: 16px;
-  border-radius: $spacing-100;
-  background-color: $white;
-  padding: 12px 20px;
-  border: 1px solid $beige-500;
-  align-items: center;
-  input {
-    border: 0;
-    outline: 0;
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-    color: $beige-500;
-    width: 248px;
-  }
-}
-
-.transactions_filter {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-}
-
-.transactions_filter-select {
-  border-radius: 8px;
-  padding: 12px 40px 12px 20px;
-  font-family: Public Sans;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 150%;
-  color: #201f24;
-  border: 1px solid var(--color-beige-500, #98908b);
-  outline: none;
-  appearance: none;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  background-image: url("../images/icon-caret-down.svg");
-  background-repeat: no-repeat;
-  background-position: right 16px center;
-  background-size: 12px;
-  cursor: pointer;
-}
-
-.transactions_filter-caption {
-  color: $grey-500;
-  font-size: $text-preset-4;
-  line-height: $line-height-150;
-  margin-right: 12px;
-}
-
-.transactions_header {
-  display: grid;
-  grid-template-columns: 3fr 1fr 1fr 2fr;
-  padding: 12px 16px;
-  border-bottom: 1px solid $grey-100;
-  span {
-    color: $grey-500;
-    font-size: $text-preset-5;
-    line-height: $line-height-150;
-  }
-}
-
-.transactions_header-amount {
-  text-align: right;
-}
-
-.transaction {
-  padding: 16px;
-  display: grid;
-  grid-template-columns: 3fr 1fr 1fr 2fr;
-  align-items: center;
-  border-bottom: solid 1px $grey-100;
-}
-
-.transaction-user {
-  display: flex;
-  gap: 16px;
-  align-items: center;
-  img {
-    height: 40px;
-    width: 40px;
-    border-radius: 50%;
-  }
-  span {
-    font-size: $text-preset-4;
-    line-height: $line-height-150;
-    font-weight: 700;
-    color: $grey-900;
-  }
-}
-
-.transaction-category,
-.transaction-amount,
-.transaction-date {
-  color: $grey-500;
-  font-size: $text-preset-5;
-  line-height: $line-height-150;
-}
-
-.transaction-amount {
-  text-align: right;
-  color: $green;
-  font-size: $text-preset-4;
-  font-weight: 700;
-  line-height: $line-height-150;
-}
-
-.active-transaction {
-  color: $green;
-}
-
-.passive-transaction {
-  color: $grey-900;
-}
