--- src/scss/base/_resets.scss
@@ -1,65 +1,0 @@
-*,
-*::after,
-*::before {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-html,
-body,
-#root {
-  height: 100%;
-}
-
-#root {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-}
-
-html {
-  font-size: 62.5%;
-  scroll-behavior: smooth;
-}
-
-ul,
-ol {
-  list-style: none;
-}
-
-a {
-  text-decoration: none;
-  display: inline-block;
-}
-
-img {
-  display: flex;
-  object-fit: cover;
-}
-
-body {
-  background-color: $beige-100;
-  font-size: $text-preset-4;
-  font-family: $global-font-family;
-}
-
-.container {
-  width: 100%;
-  margin-right: auto;
-  margin-left: auto;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-main {
-  flex-grow: 1;
-  overflow-y: auto;
-}
-
-.apexcharts-legend {
-  display: none !important;
-}
