--- src/scss/utils/_variables.scss
@@ -1,56 +1,0 @@
-// Colors
-$beige-500: #98908b;
-$beige-100: #f8f4f0;
-
-$grey-900: #201f24;
-$grey-500: #696868;
-$grey-300: #b3b3b3;
-$grey-100: #f2f2f2;
-
-$green: #277c78;
-$yellow: #f2cdac;
-$cyan: #82c9d7;
-$navy: #626070;
-$red: #c94736;
-$purple: #826cb0;
-
-// Other Colors
-$purple-alt: #af81ba;
-$turquoise: #597c7c;
-$brown: #93674f;
-$magenta: #934f6f;
-$blue: #3f82b2;
-$navy-grey: #97a0ac;
-$army-green: #7f9161;
-$gold: #cab361;
-$orange: #be6c49;
-
-// White
-$white: #ffffff;
-
-// Typography Sizes
-$text-preset-1: 3.2rem; // 32px
-$text-preset-2: 2rem; // 20px
-$text-preset-3: 1.6rem; // 16px
-$text-preset-4: 1.4rem; // 14px
-$text-preset-5: 1.2rem; // 12px
-
-// Line Heights
-$line-height-100: 1; // 100%
-$line-height-120: 1.2; // 120%
-$line-height-150: 1.5; // 150%
-
-// Letter Spacing
-$letter-spacing-0: 0rem; // 0px
-
-// Spacing Scale
-$spacing-500: 4rem; // 40px
-$spacing-400: 3.2rem; // 32px
-$spacing-300: 2.4rem; // 24px
-$spacing-250: 2rem; // 20px
-$spacing-200: 1.6rem; // 16px
-$spacing-150: 1.2rem; // 12px
-$spacing-100: 0.8rem; // 8px
-$spacing-50: 0.4rem; // 4px
-
-$global-font-family: "Public Sans", sans-serif;
