--- vercel.json
@@ -1,1 +1,0 @@
-{ "rewrites": [{ "source": "/(.*)", "destination": "/" }] }
