--- vite.config.js
@@ -1,14 +1,0 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-
-// https://vite.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  css: {
-    preprocessorOptions: {
-      scss: {
-        additionalData: `@import "./src/scss/utils/variables.scss";`,
-      },
-    },
-  },
-});
